Centralise setup flow
