+++ music_assistant/providers/telmore/setup_flow.py
@@ -1,37 +1,5 @@
 """Setup flow for the Telmore Musik provider."""
 
-from __future__ import annotations
+from music_assistant.providers.music247e.setup_flow import run_setup
 
-from dataclasses import replace
-from typing import TYPE_CHECKING
-
-from music_assistant_models.config_entries import ConfigEntry
-from music_assistant_models.enums import ConfigEntryType
-
-from music_assistant.constants import CONF_PASSWORD, CONF_USERNAME
-from music_assistant.models.setup_flow import SetupFlowError
-
-if TYPE_CHECKING:
-    from music_assistant.models.setup_flow import SetupSession
-
-_ENTRIES = (
-    ConfigEntry(key=CONF_USERNAME, type=ConfigEntryType.STRING, required=True),
-    ConfigEntry(key=CONF_PASSWORD, type=ConfigEntryType.SECURE_STRING, required=True),
-)
-
-
-async def run_setup(session: SetupSession) -> None:
-    """Run the setup flow: collect the credentials and create the provider."""
-    errors: dict[str, str] | None = None
-    setup_data = dict(session.context.setup_data)
-    while True:
-        entries = [
-            replace(entry, value=setup_data.get(entry.key, entry.value)) for entry in _ENTRIES
-        ]
-        submitted = await session.form(entries, step_id="user", errors=errors, last_step=True)
-        setup_data.update(submitted)
-        try:
-            await session.finish(setup_data)
-            return
-        except SetupFlowError as err:
-            errors = {"base": err.translation_key or str(err)}
+__all__ = ["run_setup"]
